# microservices/swagger-docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Microservices API
  description: |
    Comprehensive API documentation for Codedao microservices platform.

    ## Authentication
    Most endpoints require JWT authentication. To authenticate:
    1. Login using `POST /auth/login` 
    2. Copy the `auth_token` from response
    3. Add `Authorization: Bearer {token}` header to subsequent requests

    ## Services
    - **API Gateway** (Port 8000): Main entry point
    - **User Management** (Port 5001): Users and authentication 
    - **Exercises Service** (Port 5002): Programming exercises
    - **Scores Service** (Port 5003): User scores and results

  version: 1.0.0
  contact:
    name: Development Team
    email: dev@gmail.com
servers:
  - url: http://localhost:8000
    description: API Gateway (Recommended)
  - url: http://localhost:5001
    description: User Management Service (Direct)
  - url: http://localhost:5002
    description: Exercises Service (Direct)
  - url: http://localhost:5003
    description: Scores Service (Direct)
...
paths:
...
  /scores/:
    get:
      tags:
        - Scores
      summary: Get all scores
      description: Retrieve list of all scores
      responses:
        "200":
          description: Scores retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      scores:
                        type: array
                        items:
                          $ref: "#/components/schemas/Score"

    post:
      tags:
        - Scores
      summary: Create new score
      description: Submit a new score for an exercise
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - exercise_id
              properties:
                exercise_id:
                  type: integer
                  example: 1
                answer:
                  type: string
                  example: "def hello():\n    return 'Hello World'"
                results:
                  type: object
                  example: { "passed": true, "score": 100 }
                user_results:
                  type: object
                  example: { "time_taken": 30 }
      responses:
        "201":
          description: Score created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: New score was added!
                  data:
                    $ref: "#/components/schemas/Score"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scores/user:
    get:
      tags:
        - Scores
      summary: Get current user scores
      description: Retrieve scores for the authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User scores retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      scores:
                        type: array
                        items:
                          $ref: "#/components/schemas/Score"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scores/user/{scoreId}:
    get:
      tags:
        - Scores
      summary: Get user score by ID
      description: Retrieve specific score for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: scoreId
          in: path
          required: true
          description: Score ID
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Score retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: "#/components/schemas/Score"
        "404":
          description: Score not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scores/{exerciseId}:
    put:
      tags:
        - Scores
      summary: Update score by exercise ID
      description: Update an existing score for a specific exercise
      security:
        - BearerAuth: []
      parameters:
        - name: exerciseId
          in: path
          required: true
          description: Exercise ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answer:
                  type: string
                  example: "def hello():\n    return 'Hello Updated World'"
                results:
                  type: object
                  example: { "passed": true, "score": 95 }
                user_results:
                  type: object
                  example: { "time_taken": 25 }
      responses:
        "200":
          description: Score updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Score was updated!
                  data:
                    $ref: "#/components/schemas/Score"
        "400":
          description: Score not found or invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Score:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        exercise_id:
          type: integer
          example: 1
        answer:
          type: string
          example: "def hello():\n    return 'Hello World'"
        results:
          type: object
          example: { "passed": true, "score": 100 }
        user_results:
          type: object
          example: { "time_taken": 30 }
        created_at:
          type: string
          format: date-time
          example: "2025-09-29T19:10:00.309037"
        updated_at:
          type: string
          format: date-time
          example: "2025-09-29T19:10:00.309047"
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: fail
        message:
          type: string
          example: Error description